+++ src/nativeresources/.gitmirror
@@ -1,0 +1,1 @@
+Only contents of this folder, excluding subfolders, will be mirrored by the Git-TFS Mirror.
