+++ src/nativeresources/CMakeLists.txt
@@ -1,0 +1,9 @@
+project(nativeresourcestring)
+
+add_compile_options(-fPIC)
+
+add_library(nativeresourcestring
+  STATIC
+  resourcestring.cpp
+)
+
