+++ src/nativeresources/rctocpp.awk
@@ -57,12 +57,13 @@ function writestringentry(id, str)
 }
 
 # Write file header and begin the array we will populate with the resources.
-function writeheader()
+function writeheader(arrayName, tableName)
 {
     print "// This code was generated by rctocpp.awk and is not meant to be modified manually."
     print "#include <resourcestring.h>";
     print "";
-    print "const NativeStringResource nativeStringResources[] = {";
+    print "extern const NativeStringResourceTable " tableName ";";
+    print "const NativeStringResource " arrayName "[] = {";
 }
 
 # Write file footer
@@ -71,10 +72,13 @@ function writeheader()
 # now write all the entries we received so far when this is called.
 # After we have written all the entries, we close the array and add a 
 # constant for the size of the array for convenience.
-function writefooter()
+function writefooter(arrayName, tableName)
 {
     writesortedentries();
     print "};";
     print "";
-    print "const int NUMBER_OF_NATIVE_STRING_RESOURCES = " numEntries ";";
+
+    print "const NativeStringResourceTable " tableName " = {";
+    print numEntries ",";
+    print arrayName "};";
 }
