+++ src/utilcode/ccomprc.cpp
@@ -10,7 +10,11 @@
 #include "ndpversion.h"
 
 #include "../dlls/mscorrc/resource.h"
-#include "../dlls/mscorrc/resourcestring.h"
+#ifdef FEATURE_PAL
+#include "resourcestring.h"
+#define NATIVE_STRING_RESOURCE_NAME mscorrc_debug
+DECLARE_NATIVE_STRING_RESOURCE_TABLE(NATIVE_STRING_RESOURCE_NAME);
+#endif
 #include "sstring.h"
 #include "stringarraylist.h"
 
@@ -697,21 +701,6 @@ HRESULT CCompRC::LoadString(ResourceCategory eCategory, UINT iResourceID, __out_
     return LoadString(eCategory, langId, iResourceID, szBuffer, iMax, pcwchUsed);
 }
 
-// Used for comparing NativeStringResource elements by ID.
-int CompareNativeStringResources(const void *a, const void *b)
-{
-    unsigned int resourceIdA = ((NativeStringResource*)a)->resourceId;
-    unsigned int resourceIdB = ((NativeStringResource*)b)->resourceId;
-
-    if (resourceIdA < resourceIdB)
-        return -1;
-
-    if (resourceIdA == resourceIdB)
-        return 0;
-
-    return 1;
-}
-
 HRESULT CCompRC::LoadString(ResourceCategory eCategory, LocaleID langId, UINT iResourceID, __out_ecount(iMax) LPWSTR szBuffer, int iMax, int *pcwchUsed)
 {
     CONTRACTL
@@ -875,40 +864,8 @@ HRESULT CCompRC::LoadString(ResourceCategory eCategory, LocaleID langId, UINT iR
 
     return hr;
 #else // !FEATURE_PAL
-    int len = 0;
-    if (szBuffer && iMax)
-    {
-        // Search the sorted set of resources for the ID we're interested in.
-        NativeStringResource searchEntry = {iResourceID, NULL};
-        NativeStringResource *resourceEntry = (NativeStringResource*)bsearch(
-            &searchEntry,
-            nativeStringResources,
-            NUMBER_OF_NATIVE_STRING_RESOURCES,
-            sizeof(NativeStringResource),
-            CompareNativeStringResources);
-
-        if (resourceEntry != NULL)
-        {
-            len = PAL_GetResourceString(m_pResourceDomain, resourceEntry->resourceString, szBuffer, iMax);
-        }
-        else
-        {
-            // The resource ID wasn't found in our array. Fall back on returning the ID as a string.
-            len = _snwprintf(szBuffer, iMax - 1, W("[Undefined resource string ID:0x%X]"), iResourceID);
-            if ((len < 0) || (len == (iMax - 1)))
-            {
-                // Add string terminator if the result of _snwprintf didn't fit the buffer.
-                szBuffer[iMax - 1] = W('\0');
-                len = iMax - 1;
-            }
-        }
-    }
-
-    if (pcwchUsed)
-    {
-        *pcwchUsed = len;
-    }
-
+    LoadNativeStringResource(NATIVE_STRING_RESOURCE_TABLE(NATIVE_STRING_RESOURCE_NAME), iResourceID,
+      szBuffer, iMax, pcwchUsed);
     return S_OK;
 #endif // !FEATURE_PAL
 }
