+++ src/utilcode/dac/CMakeLists.txt
@@ -6,6 +6,7 @@ list(APPEND UTILCODE_SOURCES ../hostimpl.cpp)
 
 if(CLR_CMAKE_PLATFORM_UNIX)
     add_library(utilcode_dac STATIC ${UTILCODE_SOURCES})
+    target_link_libraries(utilcode_dac nativeresourcestring)
     add_dependencies(utilcode_dac coreclrpal)
 else()
     add_precompiled_header(stdafx.h ../stdafx.cpp UTILCODE_SOURCES)
