+++ src/utilcode/dyncrt/CMakeLists.txt
@@ -1,6 +1,7 @@
 
 if(CLR_CMAKE_PLATFORM_UNIX)
     add_library(utilcode STATIC ${UTILCODE_SOURCES})
+    target_link_libraries(utilcode nativeresourcestring)
     add_dependencies(utilcode coreclrpal)
 else()
     add_precompiled_header(stdafx.h ../stdafx.cpp UTILCODE_SOURCES)
